Requesting a game half works!
The user can request a game by selecting a user in the list box on their screen. The requested user will recieve a pop up to notify them that someone requested a game with them. They can reject or accept this request.

I need to finish the handeling for the accept/reject cases.