Server shutdown works
When the server unexpectadly shutsdown the clients will close and notify the user. Note: Rub the client from the debug folder, not Visual Studio. If you run it from Visual Studio it will complain about unhandled exceptions when the server closes. It works if you don't run it from visual studio.